requirements.txt
apache-airflow>=3.0.0,<3.1.0
kafka-python==2.0.2
requests==2.31.0
pandas==2.1.4
python-dateutil==2.8.2
